Bug fix/fix batch annotations (#124)
* web batching

* web batching

* test fix

* test fix

* test fix

+++ src/main/java/eu/dissco/backend/domain/MasJobRecordFull.java
@@ -17,7 +17,7 @@ public record MasJobRecordFull(
     Instant timeCompleted,
     JsonNode annotations,
     boolean batchingRequested,
-    Long timeToLive,
+    Integer timeToLive,
     ErrorCode errorCode) {
 
 }
